// src/enum/sys/routes.enum.ts
export enum AppRoutes {
    homepage = '',
    myProfile = 'MyProfile',
    myBaskets = 'MyBaskets',
    about = 'About',
    siteMap = 'SiteMap',

    createOrder = 'CreateOrder',

    incomingOrders = 'IncomingOrders',
    outgoingOrders = 'OutgoingOrders',
    incomingOrderReturns = 'IncomingOrderReturns',
    outgoingOrderReturns = 'OutgoingOrderReturns',

    myOrders = 'MyOrders',
    myReturns = 'MyReturns',

    login = 'Login',
    register = 'Register',
    forgotPassword = 'ForgotPassword',

    generalNew = 'New',
    generalFind = 'Find',
    generalRetail = 'Retail',
    generalConnections = 'Connections',
    generalTable = 'Table',

    accounting = 'Accounting',
    definitionsMoneyAccount = 'Wallets',
    definitionsInventory = 'Inventory',
    definitionsAddress = 'Address',

    activitiesFiche = 'FicheActivities',
    activitiesMoney = 'WalletActivities',
    activitiesStock = 'StockActivities'
}